+++ contracts/os/andromeda-vfs/src/execute.rs
@@ -155,8 +155,6 @@ pub fn register_user(
     username: String,
     address: Option<Addr>,
 ) -> Result<Response, ContractError> {
-    // #[cfg(not(test))]
-    // ensure!(false, ContractError::TemporarilyDisabled {});
     ensure!(
         username.len() as u64 <= MAX_USERNAME_LENGTH,
         ContractError::InvalidUsername {
@@ -168,7 +166,9 @@ pub fn register_user(
     let username = username.to_lowercase();
     let kernel = &ADOContract::default().get_kernel_address(env.deps.storage)?;
     let is_registration_enabled =
-        AOSQuerier::get_env_variable::<bool>(&env.deps.querier, kernel, "username_registration")?
+        AOSQuerier::get_env_variable::<String>(&env.deps.querier, kernel, "username_registration")?
+            .unwrap_or("false".to_string())
+            .parse::<bool>()
             .unwrap_or(false);
     // Can only register username directly on Andromeda chain
     ensure!(